refactor VideoBenchmark
