+++ vlmeval/dataset/utils.py
@@ -1,0 +1,1 @@
+from ..smp import *
